* une option -dont-clean pour qu'on puisse inspecter les fichiers generes * Box.shift est implemente directement comme fonction recursive plutot qu'un   appel a Box.transform   ce qui fait que Num.maxn est appele 10 a 20 fois moins souvent, ou plus si   le client utilise bcp d'alignements

+++ metapost.ml
@@ -318,21 +318,22 @@ let print_latex_error s =
               "latex -interaction=nonstopmode mpxerr.tex")
   end else ()
 
-let generate_aux bn ?prelude ?(pdf=false) ?eps ?(verbose=false) figl =
+let generate_aux bn ?prelude ?(pdf=false) ?eps ?(verbose=false) ?(clean=true) figl =
   let f = bn ^ ".mp" in
   generate_mp f ?prelude ?eps figl;
   let s,outp = 
     Misc.call_cmd ~verbose
       (sprintf "mpost -interaction=\"nonstopmode\" %s end" f) in
   if s <> 0 then print_latex_error outp;
-  ignore (Misc.call_cmd ~verbose 
-    (Printf.sprintf 
-      "rm -f mpxerr.log mpxerr.tex mpxerr.aux mpxerr.dvi %s.mp %s.mpx %s.log"
-      bn bn bn));
+  if clean then
+    ignore (Misc.call_cmd ~verbose 
+      (Printf.sprintf 
+        "rm -f mpxerr.log mpxerr.tex mpxerr.aux mpxerr.dvi %s.mp %s.mpx %s.log"
+        bn bn bn));
   if s <> 0 then exit 1
 
-let generate bn ?prelude ?(pdf=false) ?eps ?verbose figl =
-  generate_aux bn ?prelude ~pdf ?eps ?verbose figl;
+let generate bn ?prelude ?(pdf=false) ?eps ?verbose ?clean figl =
+  generate_aux bn ?prelude ~pdf ?eps ?verbose ?clean figl;
   let suf = if pdf then ".mps" else ".1" in
   let sep = if pdf then "-" else "." in
   List.iter 
@@ -342,9 +343,9 @@ let generate bn ?prelude ?(pdf=false) ?eps ?verbose figl =
     figl
 
 
-let dump ?prelude ?(pdf=false) ?eps ?(verbose=false) bn = 
+let dump ?prelude ?(pdf=false) ?eps ?(verbose=false) ?clean bn = 
   let figl = Queue.fold (fun l (i,_,f) -> (i,f) :: l) [] figures in
-  generate_aux bn ?prelude ~pdf ?eps ~verbose figl;
+  generate_aux bn ?prelude ~pdf ?eps ~verbose ?clean figl;
   let suf = if pdf then ".mps" else ".1" in
   Queue.iter 
     (fun (i,s,_) ->
